feat: add navigation component
